% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_functions.R
\name{addDays}
\alias{addDays}
\title{Create daily exposures}
\usage{
addDays(records, min_date = NULL, max_date = NULL)
}
\arguments{
\item{records}{File containing a unique policy key with start and end dates.}

\item{min_date}{Left truncation date.}

\item{max_date}{Right truncation date.}
}
\description{
This function takes a records file and creates a day for each day the policyholder is active
}
\examples{
addDays(records)
}
