% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CSO2001ANB_3NT2T_Select}
\alias{CSO2001ANB_3NT2T_Select}
\title{Long format select SOA tables 1076-1085}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 25000 rows and 6 columns.}
\usage{
CSO2001ANB_3NT2T_Select
}
\description{
Intended to be left joined to exposure frames.
}
\keyword{datasets}
