% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exposures}
\alias{exposures}
\title{Example exposure records to demonstrate calculations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24 rows and 5 columns.}
\usage{
exposures
}
\description{
A dataset containing the output of addExposures(records).
Used to demonstrate calculations.
}
\keyword{datasets}
