% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expstudies-package.R
\docType{package}
\name{expstudies}
\alias{expstudies}
\alias{expstudies-package}
\title{expstudies: package for life experience data}
\description{
Creation of an exposure table with rows for policy-intervals from a table
with a unique policy number key and beginning and ending dates for each policy.
Methods for assigning supplemental data containing dates and policy numbers
to the corresponding interval from the created exposures table.
}
\section{expstudies functions}{

\itemize{
  \item addExposures - Creation of an exposure table with either policy years or policy months.
  \item addStart - Allocate transactions to intervals.
}
}

