% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_functions.R
\name{filter_trans}
\alias{filter_trans}
\title{Remove Transactions Not in Exposures.}
\usage{
filter_trans(exposures, trans)
}
\arguments{
\item{exposures}{Output from addExposures(records).}

\item{trans}{Information we wish to filter to be within the exposure intervals.}
}
\value{
A modified version of trans where observations have matching policy keys and exposure intervals in
exposures
}
\description{
This function first converts the output of makeExposures back into a single row for each unique policy key
with a start and end date. This is joined to the corresponding key column for the transaction
and dates falling outside the exposure interval are filtered out of the transaction.
}
