% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findStart}
\alias{findStart}
\title{Find the correct interval/key combination for start intervals.}
\usage{
findStart(exp_key, start_int, end_int, trans_key, trans_date)
}
\arguments{
\item{exp_key}{sorted key column.}

\item{start_int}{start_intervals that are are sorted within their keys}

\item{end_int}{end_intervals that are are sorted because start_int is sorted.}

\item{trans_key}{that must have a matching interval. This is guaranteed by the R code before calling findStart.}

\item{trans_date}{sorted date column.}
}
\description{
Find the correct interval/key combination for start intervals.
}
