% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{makeDays}
\alias{makeDays}
\title{Create a DataFrame with all dates in the exposure period for each key}
\usage{
makeDays(record_key, record_start, record_days, final_size)
}
\arguments{
\item{record_key}{Vector of keys from the input record}

\item{record_start}{Vector of start dates from the input record}

\item{record_days}{Vector containing the number of days in the interval for a row in a record}

\item{final_size}{Sum of days in all intervals, used to initialize the DataFrame that is returned.}
}
\description{
Create a DataFrame with all dates in the exposure period for each key
}
