% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exposure_functions.R
\name{makeRange}
\alias{makeRange}
\title{All Consecutive Number Intervals Starting at 0}
\usage{
makeRange(duration)
}
\arguments{
\item{duration}{The vector containing whole numbers that will be joined to.}
}
\value{
A data frame that can be joined to the original vector by = c("duration" = "join_by").
}
\description{
A vector of whole numbers is used as input, these whole numbers are the number of
calendar years between dates from the addExposures function. Outputs is a data frame
with two columns, "join_by", and "yrs_past_start". "join_by" is intended to be joined
to the input "duration" column so that there will be rows added for each whole
number less than or equal to the duration. This is how many rows are created in the
addExposures function.
}
