% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mortality_tables}
\alias{mortality_tables}
\title{Mortality tables for A/E analysis}
\format{An object of class \code{list} of length 6.}
\usage{
mortality_tables
}
\description{
A list containing several tidy mortality tables that can be joined to an exposure data frame.
Find more tables here - github.com/ActuarialAnalyst/LongMortalityTables. Tables are from the SOA website.
}
\keyword{datasets}
