% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{records}
\alias{records}
\title{Example policy records to demonstrate calculations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.}
\usage{
records
}
\description{
A dataset containing 2 policyholders (identified by "key") as well as their issue date
("start") their termination date ("end") as well as some additional information.
Used to demonstrate calculations.
}
\keyword{datasets}
