% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trans}
\alias{trans}
\title{Example transaction records to demonstrate calculations}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 205 rows and 3 columns.}
\usage{
trans
}
\description{
A dataset containing transactions. Contains fields for the policyholder identifier ("key"),
the transaction date ("trans_date") and the amount of the premium paid ("amt").
Used to demonstrate calculations.
}
\keyword{datasets}
