% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_functions.R
\name{addStart}
\alias{addStart}
\title{Map Transaction Rows to Exposure Rows}
\usage{
addStart(exposures, trans)
}
\arguments{
\item{exposures}{Output from addExposures(records).}

\item{trans}{Information we wish to assign to exposure intervals.}
}
\value{
Modified transaction records that corresponding to an exposure interval, the interval start date
is included.
}
\description{
This function takes in exposures and transactions and returns all transactions with a matching "key"
and "trans_date" within an exposure interval and attaches the start date of the corresponding
exposure interval. This is useful for grouping by start date and key, aggregating, and joining
to the exposure rows. In this way premium pattern analysis can be performed.
}
\examples{
addStart(exposures, trans)
}
